+++ smelter-core/src/pipeline/webrtc/whep_output/stream_media_to_peer.rs
@@ -97,7 +97,7 @@ pub async fn stream_media_to_peer(
                 };
 
                 if let Some(stream) = stream {
-                    stats_sender.bytes_sent_event(chunk.data.len() as u64, chunk.kind.into());
+                    stats_sender.bytes_sent_event(chunk.data.len(), chunk.kind.into());
                     let result =
                         send_chunk_to_peer(chunk, &stream.track, &mut stream.payloader).await;
                     if let Err(err) = result {
